Factor out block and live range search functionality
